fix(code): satisfy plugin manager lint for clickable tabs
Move tab widgets out of `__init__.py` and keep Textual event imports runtime-visible for dispatch.
